+++ src/eca/messenger.clj
@@ -7,4 +7,5 @@
 (defprotocol IMessenger
   (chat-content-received [this data])
   (tool-server-updated [this params])
-  (showMessage [this msg]))
+  (showMessage [this msg])
+  (editor-diagnostics [this uri]))